A stable SHA256 signature is now computed
In order to improve the traceability of generated JSON schemas and
examples, a stable SHA256 signature is now computed from the in-memory
representation of the source OPML file. Later, that signature is
attached to the generated JSON Schema and example files.

Added also a script which allows computing and printing the stable
SHA256 signatures from input OPML files.