A stable SHA256 signature is now computed from input OPML files
In order to improve the traceability of generated JSON schemas, a stable
SHA256 signature is now computed from the in-memory representation of
the source OPML file. Later, that signature is attached to the
generated JSON Schema files.

Added also a script which allows computing and printing the stable
SHA256 signatures from input OPML files.